+++ src/BugsnagUnity/Client.cs
@@ -17,6 +17,8 @@ namespace BugsnagUnity
 
         public ISessionTracker SessionTracking { get; }
 
+        public LastRunInfo LastRunInfo => NativeClient.GetLastRunInfo();
+
         public User User { get; }
 
         public Metadata Metadata { get; }
@@ -124,7 +126,7 @@ namespace BugsnagUnity
             }
         }
 
-        private bool IsUsingFallback()
+        public bool IsUsingFallback()
         {
             switch (Application.platform)
             {
@@ -225,6 +227,18 @@ namespace BugsnagUnity
             {
                 Middleware.Add(middleware);
             }
+        }
+
+        public void Notify(string name, string message, string stackTrace, Middleware callback)
+        {
+            var exceptions = new Exception[] { Exception.FromStringInfo(name, message, stackTrace) };
+            Notify(exceptions, HandledState.ForHandledException(), callback, LogType.Exception);
+        }
+
+        public void Notify(System.Exception exception, string stacktrace, Middleware callback)
+        {
+            var exceptions = new Exceptions(exception, stacktrace).ToArray();
+            Notify(exceptions, HandledState.ForHandledException(), callback, LogType.Exception);
         }
 
         public void Notify(System.Exception exception)
@@ -415,7 +429,18 @@ namespace BugsnagUnity
                     if (Configuration.AutoTrackSessions
                      && BackgroundStopwatch.Elapsed.TotalSeconds > AutoCaptureSessionThresholdSeconds)
                     {
-                        SessionTracking.StartSession();
+                        if (IsUsingFallback())
+                        {
+                            SessionTracking.StartSession();
+                        }
+                        else
+                        {
+                            // The android sdk is unable to listen to the unity activity lifecycle
+                            if (Application.platform.Equals(RuntimePlatform.Android))
+                            {
+                                SessionTracking.ResumeSession();
+                            }
+                        }
                     }
                     BackgroundStopwatch.Reset();
                 }
@@ -443,7 +468,7 @@ namespace BugsnagUnity
         private IEnumerator<UnityEngine.AsyncOperation> RunInitialSessionCheck()
         {
             yield return null;
-            if (Configuration.AutoTrackSessions && SessionTracking.CurrentSession == null)
+            if (IsUsingFallback() && Configuration.AutoTrackSessions && SessionTracking.CurrentSession == null)
             {
                 SessionTracking.StartSession();
             }
@@ -479,6 +504,11 @@ namespace BugsnagUnity
         {
             // Set the native property
             NativeClient.SetAutoDetectAnrs(autoDetectAnrs);
+        }
+
+        public void MarkLaunchCompleted()
+        {
+            NativeClient.MarkLaunchCompleted();
         }
     }
 }