+++ src/BugsnagUnity/Payload/App.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using UnityEngine;
 
 namespace BugsnagUnity.Payload
 {
@@ -8,10 +9,17 @@ namespace BugsnagUnity.Payload
     /// </summary>
     public class App : Dictionary<string, object>, IFilterable
     {
+
+        private IConfiguration _configuration;
+
         internal App(IConfiguration configuration)
         {
+            _configuration = configuration;
             Version = configuration.AppVersion;
             ReleaseStage = configuration.ReleaseStage;
+            this.AddToPayload("id", Application.identifier);
+            this.AddToPayload("type", GetAppType());
+            AddDuration();
         }
 
         public string Version
@@ -37,5 +45,36 @@ namespace BugsnagUnity.Payload
             get => TimeSpan.FromMilliseconds((double)this.Get("durationInForeground"));
             set => this.AddToPayload("durationInForeground", value.TotalMilliseconds);
         }
+
+        private string GetAppType()
+        {
+            if (!string.IsNullOrEmpty(_configuration.AppType))
+            {
+                return _configuration.AppType;
+            }
+            switch (Application.platform)
+            {
+                case RuntimePlatform.OSXEditor:
+                case RuntimePlatform.OSXPlayer:
+                    return "MacOS";
+                case RuntimePlatform.WindowsPlayer:
+                case RuntimePlatform.WindowsEditor:
+                    return "Windows";
+                case RuntimePlatform.LinuxPlayer:
+                case RuntimePlatform.LinuxEditor:
+                    return "Linux";
+                case RuntimePlatform.WebGLPlayer:
+                    return "WebGL";
+                default:
+                    return string.Empty;
+            }
+        }
+
+        private void AddDuration()
+        {
+            var timeSinceStartup = TimeSpan.FromSeconds(Time.realtimeSinceStartup);
+            this.AddToPayload("duration", timeSinceStartup.TotalMilliseconds);
+        }
+
     }
 }