+++ tests/UnityEngine/Resolution.cs
@@ -1,0 +1,24 @@
+﻿using System;
+namespace UnityEngine
+{
+    public struct Resolution
+    {
+		/// <summary>
+		///   <para>Resolution width in pixels.</para>
+		/// </summary>
+		public int width
+		{
+			get;
+			set;
+		}
+
+		/// <summary>
+		///   <para>Resolution height in pixels.</para>
+		/// </summary>
+		public int height
+		{
+			get;
+			set;
+		}
+	}
+}