+++ tests/UnityEngine/Screen.cs
@@ -1,0 +1,13 @@
+﻿using System;
+namespace UnityEngine
+{
+    public class Screen
+    {
+        public static float dpi { get; }
+
+        public static Resolution currentResolution
+        {
+            get;
+        }
+    }
+}